[server] Split up main module
The same codebase you know and love, but now with smaller modules!

`handle_realtime_api` is split into a separate module and struct per
async task.
